+++ ts/features/pn/store/sagas/watchPnSaga.ts
@@ -3,29 +3,16 @@ import * as pot from "@pagopa/ts-commons/lib/pot";
 import { pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/lib/Option";
 import * as E from "fp-ts/lib/Either";
-import { Channel, SagaIterator, buffers, channel } from "redux-saga";
-import {
-  call,
-  flush,
-  fork,
-  put,
-  select,
-  take,
-  takeLatest
-} from "typed-redux-saga/macro";
-import { ActionType, getType, isActionOf } from "typesafe-actions";
+import { SagaIterator } from "redux-saga";
+import { call, fork, put, select, takeLatest } from "typed-redux-saga/macro";
+import { ActionType, getType } from "typesafe-actions";
 import _ from "lodash";
-import { RptIdFromString } from "@pagopa/io-pagopa-commons/lib/pagopa";
 import { apiUrlPrefix } from "../../../../config";
 import { isPnTestEnabledSelector } from "../../../../store/reducers/persistedPreferences";
 import { SessionToken } from "../../../../types/SessionToken";
 import { getError } from "../../../../utils/errors";
 import { BackendPnClient } from "../../api/backendPn";
-import {
-  updatePaymentForMessage,
-  pnActivationUpsert,
-  cancelQueuedPaymentUpdates
-} from "../actions";
+import { pnActivationUpsert } from "../actions";
 import {
   trackPNServiceStatusChangeError,
   trackPNServiceStatusChangeSuccess
@@ -33,10 +20,7 @@ import {
 import { servicePreferenceSelector } from "../../../../store/reducers/entities/services/servicePreference";
 import { isServicePreferenceResponseSuccess } from "../../../../types/services/ServicePreferenceResponse";
 import { BackendClient } from "../../../../api/backend";
-import { commonPaymentVerificationProcedure } from "../../../../sagas/wallet/pagopaApis";
-import { Detail_v2Enum } from "../../../../../definitions/backend/PaymentProblemJson";
-
-const generatePaymentUpdateWorkerCount = () => 5;
+import { watchPaymentUpdateRequests } from "./watchPaymentUpdateRequests";
 
 function* upsertPnActivation(
   client: ReturnType<typeof BackendPnClient>,
@@ -104,89 +88,4 @@ export function* watchPnSaga(
   );
 
   yield* fork(watchPaymentUpdateRequests, getVerificaRpt);
-}
-
-function* watchPaymentUpdateRequests(
-  getVerificaRpt: ReturnType<typeof BackendClient>["getVerificaRpt"]
-) {
-  // create a channel to queue incoming requests
-  const paymentUpdateChannel = yield* call(() =>
-    channel<ActionType<typeof updatePaymentForMessage.request>>(
-      buffers.expanding()
-    )
-  );
-
-  // create n worker 'threads'
-  const paymentUpdateWorkerCount = generatePaymentUpdateWorkerCount();
-  // eslint-disable-next-line functional/no-let
-  for (let i = 0; i < paymentUpdateWorkerCount; i++) {
-    yield* fork(
-      handlePaymentUpdateRequests,
-      paymentUpdateChannel,
-      getVerificaRpt
-    );
-  }
-
-  while (true) {
-    const dequeuedChannelAction = yield* take([
-      updatePaymentForMessage.request,
-      cancelQueuedPaymentUpdates
-    ]);
-
-    if (isActionOf(updatePaymentForMessage.request, dequeuedChannelAction)) {
-      yield* put(paymentUpdateChannel, dequeuedChannelAction);
-    } else {
-      const unproccessedQueuedUpdateRequests = yield* flush(
-        paymentUpdateChannel
-      );
-      const payloads = unproccessedQueuedUpdateRequests.map(
-        updateRequest => updateRequest.payload
-      );
-      if (payloads.length > 0) {
-        yield* put(updatePaymentForMessage.cancel(payloads));
-      }
-    }
-  }
-}
-
-function* handlePaymentUpdateRequests(
-  paymentStatusChannel: Channel<
-    ActionType<typeof updatePaymentForMessage.request>
-  >,
-  getVerificaRpt: ReturnType<typeof BackendClient>["getVerificaRpt"]
-) {
-  while (true) {
-    const paymentStatusRequest = yield* take(paymentStatusChannel);
-    const messageId = paymentStatusRequest.payload.messageId;
-    const paymentId = paymentStatusRequest.payload.paymentId;
-    const pagoPARptIdEither = RptIdFromString.decode(paymentId);
-    if (E.isRight(pagoPARptIdEither)) {
-      const pagoPARptId = pagoPARptIdEither.right;
-      yield* call(
-        commonPaymentVerificationProcedure,
-        getVerificaRpt,
-        pagoPARptId,
-        paymentData =>
-          updatePaymentForMessage.success({
-            messageId,
-            paymentId,
-            paymentData
-          }),
-        details =>
-          updatePaymentForMessage.failure({
-            messageId,
-            paymentId,
-            details
-          })
-      );
-    } else {
-      yield* put(
-        updatePaymentForMessage.failure({
-          messageId,
-          paymentId,
-          details: Detail_v2Enum.GENERIC_ERROR
-        })
-      );
-    }
-  }
 }
